@@ -15,7 +15,7 @@
     "test": "test"
   },
   "scripts": {
-    "test": "node_modules/.bin/mocha --reporter=nyan "
+    "build:development": "(/bin/rm -rf dist/ && NODE_ENV=development ./node_modules/.bin/webpack --config=config/webpack.config.js --mode=development --watch)"
   },
   "repository": {
     "type": "git",
@@ -30,37 +30,43 @@
     "url": "https://github.com/Mati365/i8086.js/issues"
   },
   "homepage": "https://github.com/Mati365/i8086.js#readme",
-  "dependencies": {
-    "buffer-loader": "0.0.1",
-    "color": "^0.11.3",
-    "nw": "^0.23.6",
-    "radium": "^0.18.1",
-    "react": "^15.3.1",
-    "react-dom": "^15.3.1",
-    "react-fa": "^4.1.2"
-  },
   "devDependencies": {
-    "babel": "^6.5.2",
-    "babel-core": "^6.14.0",
-    "babel-eslint": "^6.1.2",
-    "babel-loader": "^6.2.5",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-0": "^6.5.0",
-    "binary-loader": "0.0.1",
-    "chai": "^3.5.0",
-    "css-loader": "^0.25.0",
-    "eslint-plugin-babel": "^3.3.0",
-    "extract-text-webpack-plugin": "^1.0.1",
-    "file-loader": "^0.9.0",
-    "html-webpack-plugin": "^2.22.0",
-    "json-loader": "^0.5.4",
-    "mocha": "^3.0.2",
-    "pasm": "0.0.1",
-    "raw-loader": "^0.5.1",
-    "style-loader": "^0.13.1",
-    "url-loader": "^0.5.7",
-    "webpack": "^1.13.2"
+    "@babel/core": "^7.4.3",
+    "@babel/plugin-proposal-class-properties": "^7.1.0",
+    "@babel/plugin-proposal-decorators": "^7.1.2",
+    "@babel/plugin-proposal-do-expressions": "^7.0.0",
+    "@babel/plugin-proposal-function-bind": "^7.0.0",
+    "@babel/plugin-proposal-logical-assignment-operators": "^7.0.0",
+    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.4.3",
+    "@babel/plugin-proposal-optional-chaining": "^7.0.0",
+    "@babel/plugin-proposal-pipeline-operator": "^7.0.0",
+    "@babel/plugin-transform-runtime": "^7.4.3",
+    "@babel/preset-env": "^7.4.3",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/runtime": "^7.4.3",
+    "babel-eslint": "^10.0.1",
+    "babel-loader": "^8.0.4",
+    "eslint": "^5.7.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-import-resolver-webpack": "^0.11.1",
+    "eslint-loader": "^2.1.1",
+    "eslint-plugin-import": "^2.17.2",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-react": "^7.11.1",
+    "nodemon-webpack-plugin": "^4.0.8",
+    "webpack": "^4.30.0",
+    "webpack-cli": "^3.1.2",
+    "webpack-dev-server": "^3.3.1",
+    "arraybuffer-loader": "^1.0.7",
+    "buffer-loader": "^0.0.1",
+    "html-webpack-plugin": "^3.2.0"
+  },
+  "dependencies": {
+    "ramda": "^0.26.1",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6",
+    "radium": "^0.25.2",
+    "prop-types": "^15.7.2",
+    "classnames": "^2.2.6"
   }
 }
