@@ -1,29 +1,36 @@
-const React = require('react')
-    , Radium = require('radium');
+import React from 'react';
+import PropTypes from 'prop-types';
+import Radium from 'radium';
 
+export default
 @Radium
 class Output extends React.Component {
   static propTypes = {
-    onContextInit: React.PropTypes.func.isRequired,
+    onContextInit: PropTypes.func.isRequired,
   };
 
+  canvasRef = React.createRef();
+
   componentDidMount() {
+    const {canvasRef, props} = this;
+
     /** Load terminal */
-    this.props.onContextInit(this.refs.canvas);
+    props.onContextInit(canvasRef.current);
   }
 
   render() {
+    const {canvasRef} = this;
+
     return (
-      <canvas ref='canvas' style={styles} />
+      <canvas
+        ref={canvasRef}
+        style={{
+          display: 'block',
+          margin: '0 auto',
+          paddingTop: '40px',
+          imageRendering: 'pixelated',
+        }}
+      />
     );
   }
 }
-
-const styles = {
-  display: 'block',
-  margin: '0 auto',
-  paddingTop: '40px',
-  imageRendering: 'pixelated'
-};
-
-module.exports = Output;
